测试可以跑通importUser方法，还没有对保存的值进行一一比较

+++ server/bundles/com.zizibujuan.drip.server.dao/src/com/zizibujuan/drip/server/dao/DigitalIdDao.java
@@ -16,6 +16,6 @@ public interface DigitalIdDao {
 	 * @return 数字帐号
 	 * @throws SQLException 
 	 */
-	int random(Connection con) throws SQLException;
+	Long random(Connection con) throws SQLException;
 
 }
